+++ eyeball/src/shared.rs
@@ -102,11 +102,9 @@ impl<T> Observable<T> {
     /// updated value does not equal the previous value.
     pub fn set_eq(&self, value: T)
     where
-        T: Clone + PartialEq,
+        T: PartialEq,
     {
-        Self::update_eq(self, |inner| {
-            *inner = value;
-        });
+        self.state.write().unwrap().set_eq(value);
     }
 
     /// Set the inner value to the given `value` and notify subscribers if the